Update monitoring.py
Added a discriminatory field: if the logger is instantiated from a InputContainer class, the csv is excluded from the initialization and the usage. Maybe it should be handled better with polymorphism.

+++ monitoring.py
@@ -13,7 +13,7 @@ from global_variables import *
 
 
 class DataLogger(object):
-    def __init__(self, path, filename_txt, filename_csv):
+    def __init__(self, path, filename_txt, filename_csv="none"):
         self.path = path
 
         self.filename_txt = filename_txt
@@ -23,15 +23,15 @@ class DataLogger(object):
         self.heading = self.filename_csv.split("_")[0]
         self.complete_filename_csv = self.path + "\\" + self.filename_csv
 
-        self.initialize_log_file()
-        self.initialize_csv_file()
+        self.initialize_log_txt_file()
+        self.initialize_log_csv_file()
 
-    def write_log(self, text):
+    def write_log_txt(self, text):
         with open(self.complete_filename_txt, "a") as f:
             f.write(text)
             f.close()
 
-    def initialize_log_file(self):
+    def initialize_log_txt_file(self):
         try:
             os.remove(self.complete_filename_txt)
         except FileNotFoundError:
@@ -40,27 +40,28 @@ class DataLogger(object):
             with open(self.complete_filename_txt, "w") as f:
                 f.close()
 
-    def write_csv(self, step, input_level, time_process, output_level, failure, MTTF, MTTR):
+    def write_log_csv(self, step, input_level, time_process, output_level, failure, MTTF, MTTR):
         text = step + ", " + input_level + ", " + time_process + ", " + output_level + ", " + failure + ", " + \
                MTTF + ", " + MTTR
         with open(self.complete_filename_csv, "a") as f:
             f.write(text)
             f.close()
 
-    def initialize_csv_file(self):
-        try:
-            with open(self.complete_filename_csv, "w") as f:
-                f.close()
-        except FileExistsError:
-            print('The log file already exists, cleaning up and creating a new one.')
-            os.remove(self.complete_filename_csv)
-            with open(self.complete_filename_csv, "w") as f:
-                f.close()
-        finally:
-            # TODO: Maybe instead of time process should be written the number of the processed material ... or should
-            #  be added as a column
-            with open(self.complete_filename_csv, "a") as f:
-                f.write('step [step], input ' + self.heading + ' [level], time process ' + self.heading + ' [step], '
-                        'output ' + self.heading + ' [level], failure ' + self.heading + ' [bool], '
-                        'MTTF ' + self.heading + ' [step], MTTR ' + self.heading + ' [step]')
-                f.close()
+    def initialize_log_csv_file(self):
+        if self.filename_csv != "none":
+            try:
+                with open(self.complete_filename_csv, "w") as f:
+                    f.close()
+            except FileExistsError:
+                print('The log file already exists, cleaning up and creating a new one.')
+                os.remove(self.complete_filename_csv)
+                with open(self.complete_filename_csv, "w") as f:
+                    f.close()
+            finally:
+                # TODO: Maybe instead of time process should be written the number of the processed material ...
+                #  or should be added as a column
+                with open(self.complete_filename_csv, "a") as f:
+                    f.write('step [step], input ' + self.heading + ' [level], time process ' + self.heading + ' [step], '
+                            'output ' + self.heading + ' [level], failure ' + self.heading + ' [bool], '
+                            'MTTF ' + self.heading + ' [step], MTTR ' + self.heading + ' [step]\n')
+                    f.close()
